requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
confluent-kafka
httpx
yfinance
openai
sqlalchemy
psycopg2-binary
pgvector
